Försök att få till orders. Nästa steg är att ta bort tabellerna i databasen och göra en ny migrering och update och se vad som händer.

+++ MarcusRent/Models/CarViewModel.cs
@@ -6,7 +6,7 @@
         public string Brand { get; set; }
         public string Model { get; set; }
         public int Year { get; set; }
-        public float PricePerDay { get; set; }
+        public decimal PricePerDay { get; set; }
         public List<string> ImageUrls { get; set; }
     }
 }
